Move Callback implementation to separate header file
